More Windows installer improvements (#607)
- `Uninstall-SecurityDaemon` now defaults to leaving config.yaml and
  Moby data root, unless new flags `-DeleteConfig` and `-DeleteMobyDataRoot`
  are specified.

- `Uninstall-SecurityDaemon` now deletes the edge containers instead of just
  stopping them, to ensure that they're deleted even if the Moby data root
  isn't.

- `Install-SecurityDaemon` now has a new flag `-ExistingConfig` that can be
  set to use an existing config.yaml, instead of generating a new one via
  `-Manual` or `-Dps`

- Fixed a bug in `Uninstall-SecurityDaemon` where it incorrectly parsed the
  result of `docker inspect`.

- Add comment-based help, and fix two more cases of Moby Runtime -> Moby Engine.

+++ scripts/windows/setup/IotEdgeSecurityDaemon.ps1
@@ -33,41 +33,91 @@ enum ContainerOs {
     Windows
 }
 
+<#
+.SYNOPSIS
+
+Installs the IoT Edge Security Daemon and its dependencies.
+
+
+.INPUTS
+
+None
+
+
+.OUTPUTS
+
+None
+
+
+.EXAMPLE
+
+PS> Install-SecurityDaemon -Manual -DeviceConnectionString $deviceConnectionString -ContainerOs Windows
+
+
+.EXAMPLE
+
+PS> Install-SecurityDaemon -Dps -ScopeId $scopeId -RegistrationId $registrationId -ContainerOs Windows
+
+
+.EXAMPLE
+
+PS> Install-SecurityDaemon -ExistingConfig -ContainerOs Windows
+#>
 function Install-SecurityDaemon {
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName = 'Manual')]
     param (
+        # Specified the daemon will be configured manually, using a device connection string.
         [Parameter(ParameterSetName = 'Manual')]
         [Switch] $Manual,
 
+        # Specified the daemon will be configured using DPS, using a scope ID and registration ID.
         [Parameter(ParameterSetName = 'DPS')]
         [Switch] $Dps,
 
+        # Specified the daemon will be configured to use an existing config.yaml
+        [Parameter(ParameterSetName = 'ExistingConfig')]
+        [Switch] $ExistingConfig,
+
+        # The device connection string.
         [Parameter(Mandatory = $true, ParameterSetName = 'Manual')]
         [String] $DeviceConnectionString,
 
+        # The DPS scope ID.
         [Parameter(Mandatory = $true, ParameterSetName = 'DPS')]
         [String] $ScopeId,
 
+        # The DPS registration ID.
         [Parameter(Mandatory = $true, ParameterSetName = 'DPS')]
         [String] $RegistrationId,
 
+        # The base OS of all the containers that will be run on this device via the security daemon.
+        #
+        # If set to Linux, a separate installation of Docker for Windows is expected.
+        #
+        # If set to Windows, the Moby Engine will be installed automatically. This will not affect any existing installation of Docker for Windows.
         [ContainerOs] $ContainerOs = 'Linux',
 
         # Proxy URI used for all Invoke-WebRequest calls. To specify other proxy-related options like -ProxyCredential, see -InvokeWebRequestParameters
         [Uri] $Proxy,
 
-        # If set to a directory, the installer prefers to use iotedged zip, moby Engine zip, moby CLI zip and VC Runtime MSI files from inside this directory
+        # If set to a directory, the installer prefers to use iotedged zip, Moby Engine zip, Moby CLI zip and VC Runtime MSI files from inside this directory
         # over downloading them from the internet. Thus placing all four files in this directory can be used to have a completely offline install,
         # or a specific subset can be placed to override the online versions of those specific components.
         [String] $OfflineInstallationPath,
 
-        # IoT Edge Agent image to pull
+        # IoT Edge Agent image to pull for the initial configuration.
+        [Parameter(ParameterSetName = 'Manual')]
+        [Parameter(ParameterSetName = 'DPS')]
         [String] $AgentImage,
 
-        # Username to pull IoT Edge Agent image
+        # Username used to access the container registry and pull the IoT Edge Agent image.
+        [Parameter(ParameterSetName = 'Manual')]
+        [Parameter(ParameterSetName = 'DPS')]
         [String] $Username,
 
-        # Password to pull IoT Edge Agent image
+        # Password used to access the container registry and pull the IoT Edge Agent image.
+        [Parameter(ParameterSetName = 'Manual')]
+        [Parameter(ParameterSetName = 'DPS')]
         [SecureString] $Password,
 
         # Splatted into every Invoke-WebRequest invocation. Can be used to set extra options.
@@ -75,10 +125,11 @@ function Install-SecurityDaemon {
         # If -Proxy is also specified, it overrides the `-Proxy` key set in this hashtable, if any.
         #
         # Example:
+        #
         #     Install-SecurityDaemon -InvokeWebRequestParameters @{ '-Proxy' = 'http://localhost:8888'; '-ProxyCredential' = (Get-Credential).GetNetworkCredential() }
         [HashTable] $InvokeWebRequestParameters,
 
-        # Don't install the Moby CLI (docker.exe) to $MobyInstallDirectory. Only takes effect for `-ContainerOs Windows`
+        # Don't install the Moby CLI (docker.exe) to $MobyInstallDirectory. Only takes effect for -ContainerOs 'Windows'
         [Switch] $SkipMobyCli,
 
         # Local path to iotedged zip file. Only kept for backward compatibility. Prefer to set -OfflineInstallationPath instead.
@@ -166,18 +217,40 @@ function Install-SecurityDaemon {
             [System.Net.ServicePointManager]::SecurityProtocol -bor [System.Net.SecurityProtocolType]::Tls12
     }
 
+    if ($ExistingConfig -and (-not (Test-Path "$EdgeInstallDirectory\config.yaml"))) {
+        Write-HostRed
+        Write-HostRed "$EdgeInstallDirectory\config.yaml was not found."
+        return
+    }
+
+    if ((-not $ExistingConfig) -and (Test-Path "$EdgeInstallDirectory\config.yaml")) {
+        Write-HostRed
+        Write-HostRed "$EdgeInstallDirectory\config.yaml already exists."
+        Write-HostRed (
+            'Delete it using `Uninstall-SecurityDaemon -Force -DeleteConfig` and then re-run `Install-SecurityDaemon`, ' +
+            'or run `Install-SecurityDaemon -ExistingConfig` to use the existing config.yaml')
+        return
+    }
+
     # Download
     Get-SecurityDaemon
     Get-VcRuntime
 
-    # Update config.yaml
-    Set-ProvisioningMode
-    Set-AgentImage
-    Set-Hostname
-    if ($ContainerOs -eq 'Linux') {
-        Set-GatewayAddress
+    # config.yaml
+    if ($ExistingConfig) {
+        Write-HostGreen 'Using existing config.yaml'
     }
-    Set-MobyEngineParameters
+    else {
+        Write-Host 'Generating config.yaml...'
+
+        Set-ProvisioningMode
+        Set-AgentImage
+        Set-Hostname
+        if ($ContainerOs -eq 'Linux') {
+            Set-GatewayAddress
+        }
+        Set-MobyEngineParameters
+    }
 
     # Register services
     Set-SystemPath
@@ -195,9 +268,53 @@ function Install-SecurityDaemon {
     Write-HostGreen '    Format-Table -AutoSize -Wrap'
 }
 
+<#
+.SYNOPSIS
+
+Uninstalls the IoT Edge Security Daemon and its dependencies.
+
+
+.DESCRIPTION
+
+By default this commandlet does not delete the config.yaml and the Moby Engine data root (for -ContainerOs 'Windows' installs),
+so that you can update the daemon using `Install-SecurityDaemon -ExistingConfig`. To delete these as well, specify the
+-DeleteConfig and -DeleteMobyDataRoot flags.
+
+
+.INPUTS
+
+None
+
+
+.OUTPUTS
+
+None
+
+
+.EXAMPLE
+
+PS> Uninstall-SecurityDaemon
+
+
+.EXAMPLE
+
+PS> Uninstall-SecurityDaemon -DeleteConfig -DeleteMobyDataRoot
+
+
+.EXAMPLE
+
+PS> Uninstall-SecurityDaemon -Force
+#>
 function Uninstall-SecurityDaemon {
     [CmdletBinding()]
     param (
+        # If set, the config.yaml will also be deleted.
+        [Switch] $DeleteConfig,
+
+        # If set, the Moby Engine data root will also be deleted. This directory is only created for -ContainerOs 'Windows' installs.
+        [Switch] $DeleteMobyDataRoot,
+
+        # Forces the uninstallation in case the previous install was only partially successful.
         [Switch] $Force
     )
 
@@ -215,7 +332,6 @@ function Uninstall-SecurityDaemon {
     $ContainerOs = Get-ContainerOs
 
     Uninstall-Services
-    Stop-IotEdgeContainers
     $success = Remove-SecurityDaemonResources
     Reset-SystemPath
 
@@ -297,11 +413,11 @@ function Get-WindowsBuild {
 }
 
 function Test-EdgeAlreadyInstalled {
-    (Get-Service $EdgeServiceName -ErrorAction SilentlyContinue) -or (Test-Path -Path $EdgeInstallDirectory)
+    (Get-Service $EdgeServiceName -ErrorAction SilentlyContinue) -or (Test-Path -Path "$EdgeInstallDirectory\iotedged.exe")
 }
 
 function Test-MobyAlreadyInstalled {
-    (Get-Service $MobyServiceName -ErrorAction SilentlyContinue) -or (Test-Path -Path $MobyInstallDirectory) -or (Test-Path -Path $MobyDataRootDirectory)
+    (Get-Service $MobyServiceName -ErrorAction SilentlyContinue) -or (Test-Path -Path $MobyInstallDirectory)
 }
 
 function Test-AgentRegistryArgs {
@@ -361,7 +477,7 @@ function Get-SecurityDaemon {
                     -Delete ([ref] $deleteMobyEngineArchive)
             Expand-Archive $mobyEngineArchivePath $MobyInstallDirectory -Force
 
-            New-Item -Type Directory $MobyDataRootDirectory | Out-Null
+            New-Item -Type Directory $MobyDataRootDirectory -Force | Out-Null
             Remove-BuiltinWritePermissions $MobyDataRootDirectory
 
             if (-not ($SkipMobyCli)) {
@@ -396,13 +512,23 @@ function Get-SecurityDaemon {
         }
 
         if ((Get-Item $edgeArchivePath).PSIsContainer) {
-            New-Item -Type Directory $EdgeInstallDirectory | Out-Null
-            Copy-Item "$edgeArchivePath\*" $EdgeInstallDirectory -Force
+            New-Item -Type Directory $EdgeInstallDirectory -Force | Out-Null
+            if ($ExistingConfig) {
+                Copy-Item "$edgeArchivePath\*" $EdgeInstallDirectory -Force -Recurse -Exclude 'config.yaml'
+            }
+            else {
+                Copy-Item "$edgeArchivePath\*" $EdgeInstallDirectory -Force -Recurse
+            }
         }
         else {
             New-Item -Type Directory $EdgeInstallDirectory | Out-Null
             Expand-Archive $edgeArchivePath $EdgeInstallDirectory -Force
-            Copy-Item "$EdgeInstallDirectory\iotedged-windows\*" $EdgeInstallDirectory -Force -Recurse
+            if ($ExistingConfig) {
+                Copy-Item "$EdgeInstallDirectory\iotedged-windows\*" $EdgeInstallDirectory -Force -Recurse -Exclude 'config.yaml'
+            }
+            else {
+                Copy-Item "$EdgeInstallDirectory\iotedged-windows\*" $EdgeInstallDirectory -Force -Recurse
+            }
         }
 
         Remove-BuiltinWritePermissions $EdgeInstallDirectory
@@ -476,7 +602,17 @@ function Remove-SecurityDaemonResources {
     Remove-Item $logKey -ErrorAction SilentlyContinue -ErrorVariable cmdErr
     Write-Verbose "$(if ($?) { "Deleted registry key '$logKey'" } else { $cmdErr })"
 
-    Remove-Item -Recurse $EdgeInstallDirectory -ErrorAction SilentlyContinue -ErrorVariable cmdErr
+    Write-Host "Deleting install directory '$EdgeInstallDirectory'..."
+    if (-not $DeleteConfig) {
+        Write-Host "Not deleting config.yaml since -DeleteConfig was not specified."
+    }
+
+    if ($DeleteConfig) {
+        Remove-Item -Recurse $EdgeInstallDirectory -ErrorAction SilentlyContinue -ErrorVariable cmdErr
+    }
+    else {
+        Remove-Item -Recurse $EdgeInstallDirectory -Exclude 'config.yaml' -ErrorAction SilentlyContinue -ErrorVariable cmdErr
+    }
     if ($?) {
         Write-Verbose "Deleted install directory '$EdgeInstallDirectory'"
     }
@@ -505,32 +641,37 @@ function Remove-SecurityDaemonResources {
         Write-Verbose "$cmdErr"
     }
 
-    if (Test-Path $MobyDataRootDirectory) {
-        try {
-            Write-Host "Deleting Moby data root directory '$MobyDataRootDirectory'..."
+    if ($DeleteMobyDataRoot) {
+        if (Test-Path $MobyDataRootDirectory) {
+            try {
+                Write-Host "Deleting Moby data root directory '$MobyDataRootDirectory'..."
 
-            # Removing `$MobyDataRootDirectory` is tricky. Windows base images contain files owned by TrustedInstaller, etc
-            # Deleting them is a three-step process:
-            #
-            # 1. Take ownership of all files
-            Invoke-Native "takeown /r /skipsl /f ""$MobyDataRootDirectory"""
+                # Removing `$MobyDataRootDirectory` is tricky. Windows base images contain files owned by TrustedInstaller, etc
+                # Deleting them is a three-step process:
+                #
+                # 1. Take ownership of all files
+                Invoke-Native "takeown /r /skipsl /f ""$MobyDataRootDirectory"""
 
-            # 2. Reset their ACLs so that they inherit from their container
-            Invoke-Native "icacls ""$MobyDataRootDirectory"" /reset /t /l /q /c"
+                # 2. Reset their ACLs so that they inherit from their container
+                Invoke-Native "icacls ""$MobyDataRootDirectory"" /reset /t /l /q /c"
 
-            # 3. Use cmd's `rd` rather than `Remove-Item` since the latter gets tripped up by reparse points, etc.
-            #    Prepend the path with `\\?\` since the layer directories have long names, so the paths usually exceed 260 characters,
-            #    and IoT Core's filesystem doesn't seem to automatically use (or even have) short names
-            Invoke-Native "rd /s /q ""\\?\$MobyDataRootDirectory"""
+                # 3. Use cmd's `rd` rather than `Remove-Item` since the latter gets tripped up by reparse points, etc.
+                #    Prepend the path with `\\?\` since the layer directories have long names, so the paths usually exceed 260 characters,
+                #    and IoT Core's filesystem doesn't seem to automatically use (or even have) short names
+                Invoke-Native "rd /s /q ""\\?\$MobyDataRootDirectory"""
 
-            Write-Verbose "Deleted Moby data root directory '$MobyDataRootDirectory'"
+                Write-Verbose "Deleted Moby data root directory '$MobyDataRootDirectory'"
+            }
+            catch {
+                Write-Verbose "$_"
+                Write-HostRed ("Could not delete Moby data root directory '$MobyDataRootDirectory'. Please reboot " +
+                    'your device and run `Uninstall-SecurityDaemon` again with `-Force`.')
+                $success = $false
+            }
         }
-        catch {
-            Write-Verbose "$_"
-            Write-HostRed ("Could not delete Moby data root directory '$MobyDataRootDirectory'. Please reboot " +
-                'your device and run `Uninstall-SecurityDaemon` again with `-Force`.')
-            $success = $false
-        }
+    }
+    else {
+        Write-Host "Not deleting Moby data root directory '$MobyDataRootDirectory' since -DeleteMobyDataRoot was not specified."
     }
 
     Remove-Item -Recurse $MobyInstallDirectory -ErrorAction SilentlyContinue -ErrorVariable cmdErr
@@ -692,24 +833,24 @@ function Uninstall-Services {
         }
     }
 
+    Remove-IotEdgeContainers
+
     if (Get-Service $MobyServiceName -ErrorAction SilentlyContinue) {
         Stop-Service -NoWait -ErrorAction SilentlyContinue -ErrorVariable cmdErr $MobyServiceName
         if ($?) {
             Start-Sleep -Seconds 7
-            Write-Verbose 'Stopped the IoT Edge Moby Runtime service'
+            Write-Verbose 'Stopped the IoT Edge Moby Engine service'
         }
         else {
             Write-Verbose "$cmdErr"
         }
 
         if (Invoke-Native "sc.exe delete ""$MobyServiceName""" -ErrorAction SilentlyContinue) {
-            Write-Verbose 'Removed IoT Edge Moby Runtime service subkey from the registry'
+            Write-Verbose 'Removed IoT Edge Moby Engine service subkey from the registry'
         }
     }
 }
 
-# The daemon listens using UDS by default, so these firewall rules are no longer created.
-# However Uninstall-SecurityDaemon still calls this to remove rules created by old installs.
 function Remove-FirewallExceptions {
     Remove-NetFirewallRule -DisplayName 'iotedged allow inbound 15580,15581' -ErrorAction SilentlyContinue -ErrorVariable cmdErr
     Write-Verbose "$(if ($?) { 'Removed firewall exceptions' } else { $cmdErr })"
@@ -849,23 +990,36 @@ function Get-AgentRegistry {
     return 'index.docker.io'
 }
 
-function Stop-IotEdgeContainers {
+function Remove-IotEdgeContainers {
     $dockerExe = Get-DockerExePath
 
     if (-not (Test-IsDockerRunning 6> $null)) {
         return
     }
 
-    $runningContainersString = Invoke-Native "$dockerExe ps --format ""{{.ID}}""" -Passthru
-    [string[]] $runningContainers = $runningContainersString -split {$_ -eq "`r" -or $_ -eq "`n"} | where {$_.Length -gt 0}
+    $allContainersString = Invoke-Native "$dockerExe ps --all --format ""{{.ID}}""" -Passthru
+    [string[]] $allContainers = $allContainersString -split {$_ -eq "`r" -or $_ -eq "`n"} | where {$_.Length -gt 0}
 
-    foreach ($containerId in $runningContainers) {
-        $inspectString = Invoke-Native "$dockerExe inspect ""$containerId""" -Passthru | ConvertFrom-Json
+    foreach ($containerId in $allContainers) {
+        $inspectString = Invoke-Native "$dockerExe inspect ""$containerId""" -Passthru
         $inspectResult = ($inspectString | ConvertFrom-Json)[0]
 
         if ($inspectResult.Config.Labels.'net.azure-devices.edge.owner' -eq 'Microsoft.Azure.Devices.Edge.Agent') {
-            Invoke-Native "$dockerExe stop ""$containerId"""
-            Write-Verbose "Stopped container $($inspectResult.Name)"
+            if (($inspectResult.Name -eq '/edgeAgent') -or ($inspectResult.Name -eq '/edgeHub')) {
+                Invoke-Native "$dockerExe rm --force ""$containerId"""
+                Write-Verbose "Stopped and deleted container $($inspectResult.Name)"
+
+                # `.Config.Image` contains the user-provided name, but this can be a tag like `foo:latest`
+                # that doesn't necessarily point to the image that has that tag right now.
+                #
+                # So delete the image using its unique identifier, as given by `.Image`, like `sha256:1234abcd...`
+                Invoke-Native "$dockerExe image rm --force ""$($inspectResult.Image)"""
+                Write-Verbose "Deleted image of container $($inspectResult.Name) ($($inspectResult.Config.Image))"
+            }
+            else {
+                Invoke-Native "$dockerExe stop ""$containerId"""
+                Write-Verbose "Stopped container $($inspectResult.Name)"
+            }
         }
     }
 }
